Added Code of Conduct
From now on, the project is claimed by evil leftoids and, under no circumstances, can it be unclaimed.

It's been quite a while since the PolyMC drama, but I'm still impressed by how a few lines of text - simply stating that it's not okay to harass somebody - can cause a literal mental breakdown for some conservatives.